Scenario 2. working

+++ libs/icore/src/vil/update_assets.rs
@@ -15,7 +15,7 @@ pub fn update_assets(
     delta_long_id: u128,
     delta_short_id: u128,
     margin_id: u128,
-) -> Vec<u128> {
+) -> Vec<u8> {
     devil! {
         // ====================================
         // * * * (TRY) COMPUTE NEW VALUES * * *
